Add files via upload
In this project, I did a proper analysis of banking dataset, which includes two kinds of data set, customer & credit card. This dynamic dashboard provides us to know the actual data and variation in a better way. With the help of this dashboard, we can make strategic decisions based on what is working right and what is not for the business and take actions accordingly.